feat: initial configuration to use uwsgi
This sets up the securemailbox flask app to be served
by uwsgi instead of flask's built in dev server.
There are two configurations:
 - one to serve it directly
 - one to run behind nginx.
The two configurations are in the .ini files and which
runs depends on which Dockerfile is used to build the image.
The .prod is for running in conjunction with nginx.